Refaktoroitu KOMOTO lomakkeelta FormConfig pois
