+++ src/main/app/src/hooks/form.ts
@@ -12,7 +12,6 @@ import { assert } from '#/src/utils';
 import { getKielivalinta } from '#/src/utils/form/formConfigUtils';
 import getHakuFormConfig from '#/src/utils/haku/getHakuFormConfig';
 import getHakukohdeFormConfig from '#/src/utils/hakukohde/getHakukohdeFormConfig';
-import getToteutusFormConfig from '#/src/utils/toteutus/getToteutusFormConfig';
 
 import { useActions } from './useActions';
 
@@ -62,7 +61,6 @@ export function useFieldValue<T = any>(name, formNameProp?: string): T {
 }
 
 const formConfigsGettersByEntity = {
-  [ENTITY.TOTEUTUS]: getToteutusFormConfig,
   [ENTITY.HAKU]: getHakuFormConfig,
   [ENTITY.HAKUKOHDE]: getHakukohdeFormConfig,
 };
