+++ src/main/app/src/pages/toteutus/ToteutusForm/LukiolinjatSection/index.tsx
@@ -5,5 +5,3 @@ import LukionLinjatField from './LukionLinjatField';
 export const LukiolinjatSection = ({ name }) => {
   return <LukionLinjatField name={`${name}.lukiolinja`} />;
 };
-
-export default LukiolinjatSection;
