+++ src/main/app/src/pages/toteutus/ToteutusForm/TiedotSection.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 
+import _fp from 'lodash/fp';
 import { useTranslation } from 'react-i18next';
 import { Field } from 'redux-form';
 
@@ -10,7 +11,12 @@ import {
   FormFieldSwitch,
 } from '#/src/components/formFields';
 import { Box } from '#/src/components/virkailija';
-import { KOULUTUSTYYPPI } from '#/src/constants';
+import {
+  KOULUTUSTYYPIT,
+  KOULUTUSTYYPPI,
+  TUTKINTOON_JOHTAMATTOMAT_KOULUTUSTYYPIT,
+  TUTKINTOON_JOHTAVAT_AMMATILLISET_KOULUTUSTYYPIT,
+} from '#/src/constants';
 import useKoodistoOptions from '#/src/hooks/useKoodistoOptions';
 import { getTestIdProps } from '#/src/utils';
 
@@ -47,7 +53,7 @@ const LaajuusFields = ({ name, disabled }) => {
   );
 };
 
-const TiedotSection = ({ language, name, koulutustyyppi }) => {
+export const TiedotSection = ({ language, name, koulutustyyppi }) => {
   const { t } = useTranslation();
 
   const disableFieldsCopiedFromKoulutus = [
@@ -57,54 +63,99 @@ const TiedotSection = ({ language, name, koulutustyyppi }) => {
 
   return (
     <>
-      <Box mb={2} {...getTestIdProps('toteutuksenNimi')}>
-        <Field
-          name={`${name}.nimi.${language}`}
-          component={FormFieldInput}
-          disabled={disableFieldsCopiedFromKoulutus}
-          label={t('toteutuslomake.toteutuksenNimi')}
-        />
-      </Box>
+      {_fp
+        .without(
+          [
+            KOULUTUSTYYPPI.LUKIOKOULUTUS,
+            KOULUTUSTYYPPI.VALMA,
+            KOULUTUSTYYPPI.TELMA,
+            KOULUTUSTYYPPI.LUVA,
+            KOULUTUSTYYPPI.PERUSOPETUKSEN_LISAOPETUS,
+          ],
+          KOULUTUSTYYPIT
+        )
+        .includes(koulutustyyppi) && (
+        <Box mb={2} {...getTestIdProps('toteutuksenNimi')}>
+          <Field
+            name={`${name}.nimi.${language}`}
+            component={FormFieldInput}
+            disabled={disableFieldsCopiedFromKoulutus}
+            label={t('toteutuslomake.toteutuksenNimi')}
+            required
+          />
+        </Box>
+      )}
 
-      <Box mb={2}>
-        <Field
-          name={`${name}.ammatillinenPerustutkintoErityisopetuksena`}
-          component={FormFieldSwitch}
-        >
-          {t('toteutuslomake.ammatillinenPerustutkintoErityisopetuksena')}
-        </Field>
-      </Box>
+      {TUTKINTOON_JOHTAVAT_AMMATILLISET_KOULUTUSTYYPIT.includes(
+        koulutustyyppi
+      ) && (
+        <Box mb={2}>
+          <Field
+            name={`${name}.ammatillinenPerustutkintoErityisopetuksena`}
+            component={FormFieldSwitch}
+          >
+            {t('toteutuslomake.ammatillinenPerustutkintoErityisopetuksena')}
+          </Field>
+        </Box>
+      )}
+      {
+        /* TODO: name-attribuutin alkuun sectionin name. Muuten validointivirheen tullessa tätä kenttää ei osata näyttää oikein. */
+        _fp
+          .without([KOULUTUSTYYPPI.LUKIOKOULUTUS], KOULUTUSTYYPIT)
+          .includes(koulutustyyppi) && (
+          <Box mb={2} {...getTestIdProps('toteutuksenKuvaus')}>
+            <Field
+              name={`kuvaus.${language}`}
+              component={FormFieldEditor}
+              label={t('toteutuslomake.toteutuksenYleinenKuvaus')}
+            />
+          </Box>
+        )
+      }
 
-      <Box mb={2} {...getTestIdProps('toteutuksenKuvaus')}>
-        <Field
-          name={`kuvaus.${language}`}
-          component={FormFieldEditor}
-          label={t('toteutuslomake.toteutuksenYleinenKuvaus')}
-        />
-      </Box>
+      {_fp
+        .without(
+          [KOULUTUSTYYPPI.OSAAMISALA, KOULUTUSTYYPPI.TUTKINNON_OSA],
+          TUTKINTOON_JOHTAMATTOMAT_KOULUTUSTYYPIT
+        )
+        .includes(koulutustyyppi) && (
+        <>
+          <Box mb={2} {...getTestIdProps('ilmoittautumislinkki')}>
+            <Field
+              name={`${name}.ilmoittautumislinkki.${language}`}
+              component={FormFieldInput}
+              label={t('toteutuslomake.ilmoittautumislinkki')}
+            />
+          </Box>
+          <Box mb={2}>
+            <LaajuusFields
+              name={name}
+              disabled={disableFieldsCopiedFromKoulutus}
+            />
+          </Box>
+        </>
+      )}
 
-      <Box mb={2} {...getTestIdProps('ilmoittautumislinkki')}>
-        <Field
-          name={`${name}.ilmoittautumislinkki.${language}`}
-          component={FormFieldInput}
-          label={t('toteutuslomake.ilmoittautumislinkki')}
-        />
-      </Box>
-
-      <Box mb={2}>
-        <LaajuusFields name={name} disabled={disableFieldsCopiedFromKoulutus} />
-      </Box>
-
-      <Box mb={2} {...getTestIdProps('aloituspaikat')}>
-        <Field
-          name={`${name}.aloituspaikat`}
-          component={FormFieldInput}
-          label={t('toteutuslomake.aloituspaikat')}
-          type="number"
-        />
-      </Box>
+      {_fp
+        .without(
+          [
+            KOULUTUSTYYPPI.VALMA,
+            KOULUTUSTYYPPI.TELMA,
+            KOULUTUSTYYPPI.LUVA,
+            KOULUTUSTYYPPI.PERUSOPETUKSEN_LISAOPETUS,
+          ],
+          TUTKINTOON_JOHTAMATTOMAT_KOULUTUSTYYPIT
+        )
+        .includes(koulutustyyppi) && (
+        <Box mb={2} {...getTestIdProps('aloituspaikat')}>
+          <Field
+            name={`${name}.aloituspaikat`}
+            component={FormFieldInput}
+            label={t('toteutuslomake.aloituspaikat')}
+            type="number"
+          />
+        </Box>
+      )}
     </>
   );
 };
-
-export default TiedotSection;
