+++ src/main/app/src/pages/toteutus/ToteutusForm/ToteutusjaksotSection.tsx
@@ -97,5 +97,3 @@ export const ToteutusjaksotSection = ({ name, language = 'fi' }) => {
     />
   );
 };
-
-export default ToteutusjaksotSection;
